Replacing int with IntType in communicator header.

+++ model_zoo/lbann2.cpp
@@ -360,7 +360,10 @@ bool load_model_weights(std::string ckpt_dir, model * m){
     if(ret != 2) { return false; }
     closeread(fd, latest);
     if(temp_comm->am_model_master())
-      sprintf(latest, "%s/shared.model.%d.epoch.%d.step.%d/", ckpt_dir.c_str(), temp_comm->get_model_rank(), epochLast, stepLast);
+      sprintf(latest,
+              "%s/shared.model.%d.epoch.%d.step.%d/",
+              ckpt_dir.c_str(), (int) temp_comm->get_model_rank(),
+              (int) epochLast, (int) stepLast);
     temp_comm->model_broadcast(0, &(latest[0]), sizeof(latest));
   }
